renames entity to dto
