+++ pom.xml
@@ -10,7 +10,7 @@
     </parent>
     <groupId>xyz.wiraharas</groupId>
     <artifactId>wiraharas-xyz-service</artifactId>
-    <version>2.0.0</version>
+    <version>2.1.0</version>
     <name>wiraharas-xyz-service</name>
     <description>Spring Boot service</description>
