+++ src/main/java/xyz/wiraharas/controller/Controller.java
@@ -1,14 +1,14 @@
 package xyz.wiraharas.controller;
 
 import java.util.List;
-import xyz.wiraharas.entity.Cuenta;
+import xyz.wiraharas.dto.CuentaDto;
 import xyz.wiraharas.service.CuentaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
-import xyz.wiraharas.entity.Titular;
-import xyz.wiraharas.entity.User;
+import xyz.wiraharas.dto.TitularDto;
+import xyz.wiraharas.dto.UserDto;
 import xyz.wiraharas.service.TitularService;
 import xyz.wiraharas.service.UserService;
 
@@ -25,19 +25,19 @@ class Controller {
     private UserService userService;
 
     @GetMapping("/cuenta/{guid}")
-    Cuenta getCuenta(@PathVariable String guid) {
+    CuentaDto getCuenta(@PathVariable String guid) {
 
         return cuentaService.get(guid);
     }
 
     @GetMapping("/titular/{guid}")
-    Titular getTitular(@PathVariable String guid) {
+    TitularDto getTitular(@PathVariable String guid) {
 
         return titularService.get(guid);
     }
 
     @GetMapping("/usuario")
-    List<User> listAllUsers() {
+    List<UserDto> listAllUsers() {
 
         return userService.listAll();
     }
