+++ src/main/java/xyz/wiraharas/dto/CuentaDto.java
@@ -1,0 +1,28 @@
+package xyz.wiraharas.dto;
+
+public class CuentaDto {
+
+    private String rut;
+    private String number;
+
+    public CuentaDto() {
+        super();
+    }
+
+    public String getRut() {
+        return rut;
+    }
+
+    public void setRut(String rut) {
+        this.rut = rut;
+    }
+
+    public String getNumber() {
+        return number;
+    }
+
+    public void setNumber(String number) {
+        this.number = number;
+    }
+
+}
