+++ src/main/java/xyz/wiraharas/dto/TitularDto.java
@@ -1,0 +1,46 @@
+package xyz.wiraharas.dto;
+
+public class TitularDto {
+
+    private String name;
+    private String lastName;
+    private String email;
+    private String rut;
+
+    public TitularDto() {
+        super();
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getRut() {
+        return rut;
+    }
+
+    public void setRut(String rut) {
+        this.rut = rut;
+    }
+
+}
