+++ src/main/java/xyz/wiraharas/dto/UserDto.java
@@ -1,0 +1,37 @@
+package xyz.wiraharas.dto;
+
+public class UserDto {
+
+    private String name;
+    private String lastName;
+    private String email;
+
+    public UserDto() {
+        super();
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+}
