--- src/main/java/xyz/wiraharas/entity/Cuenta.java
@@ -1,57 +1,0 @@
-package xyz.wiraharas.entity;
-
-import java.util.Objects;
-
-public class Cuenta {
-
-    private String rut;
-    private String number;
-
-    public Cuenta() {
-    }
-
-    public String getRut() {
-        return rut;
-    }
-
-    public void setRut(String rut) {
-        this.rut = rut;
-    }
-
-    public String getNumber() {
-        return number;
-    }
-
-    public void setNumber(String number) {
-        this.number = number;
-    }
-
-    @Override
-    public int hashCode() {
-        int hash = 5;
-        hash = 67 * hash + Objects.hashCode(this.rut);
-        hash = 67 * hash + Objects.hashCode(this.number);
-        return hash;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj == null) {
-            return false;
-        }
-        if (getClass() != obj.getClass()) {
-            return false;
-        }
-        final Cuenta other = (Cuenta) obj;
-        if (!Objects.equals(this.number, other.number)) {
-            return false;
-        }
-        return true;
-    }
-
-    @Override
-    public String toString() {
-        return "Cuenta{" + "rut=" + rut + ", number=" + number + '}';
-    }
-
-}
