+++ src/main/java/xyz/wiraharas/service/UserService.java
@@ -1,10 +1,10 @@
 package xyz.wiraharas.service;
 
-import xyz.wiraharas.entity.User;
+import xyz.wiraharas.dto.UserDto;
 import java.util.List;
 
 public interface UserService {
 
-    public List<User> listAll();
+    List<UserDto> listAll();
 
 }
