+++ src/main/java/xyz/wiraharas/service/impl/UserServiceImpl.java
@@ -1,6 +1,6 @@
 package xyz.wiraharas.service.impl;
 
-import xyz.wiraharas.entity.User;
+import xyz.wiraharas.dto.UserDto;
 import xyz.wiraharas.service.UserService;
 import java.util.ArrayList;
 import java.util.List;
@@ -10,10 +10,9 @@ import org.springframework.stereotype.Service;
 public class UserServiceImpl implements UserService {
 
     @Override
-    public List<User> listAll() {
-        List<User> users = new ArrayList<>();
-        User user = new User();
-        user.setId(1L);
+    public List<UserDto> listAll() {
+        List<UserDto> users = new ArrayList<>();
+        UserDto user = new UserDto();
         user.setName("John");
         user.setLastName("Doe");
         user.setEmail("john.doe@wiraharas.xyz");
